Create My first code
I created this code by taking the C# course I'm taking with the Microsoft Online educational institution. It seeks to show the calculation of the students' grade, show the alphabetical grade, in addition to calculating the sum of the extra activities that add 10% to the value of the student's average.