BMI calculation
In this exercise I created the code myself from the teacher's instructions as to the purpose of the program. I used the Boolean parameters of if, else if, and else to build the code. Here it is possible to calculate the body BMI of an individual, returning the message of the situation of the same.