Polymorphism exercise
In this exercise is shown in the terminal the discount that you have on the salary of the intern, the manager and the attendant, on the food voucher and the transport voucher. The manager and the attendant have discounts different from the standard that the trainee has, so the use of polymorphism to create this code.